Draft Figure 1 Script
Addes in 01_fig1.py which plots the actograms for all the animals
required but with a script rather than powerpoint.
Previous version of fig 1 used individually created actograms which
were then arranged in powerpoint.
This version uses python to handle all the subplots.

Still requires tidying of axis and labels but mostly there